pyproject.toml
[project]
name = "fastapi-auth-server"
version = "0.2.0"
description = "Production-ready FastAPI authentication server with Google OAuth and JWT."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.24.0",
    "authlib>=1.2.0",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.0.0",
    "itsdangerous>=2.0.0",
    "httpx>=0.25.0",
    "pydantic[email]>=2.0.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0",
    "cryptography>=3.4.0",
    "python-jose[cryptography]>=3.3.0",
    "jinja2>=3.1.6",
]

[tool.uv]
# Optional: uv specific configurations

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# Google Python Style Guide enforcement tools
[tool.isort]
profile = "google"
line_length = 100

[tool.yapf]
based_on_style = "google"
column_limit = 100

[tool.pylint]
py-version = "3.10"
fail-under = 9.0
jobs = 4
persistent = true
max-line-length = 100
disable = [
    "R", # Refactoring suggestions
    "C0111", # Missing docstring (handled by our docstring conventions)
    "C0103", # Invalid name (we'll handle naming with our own patterns)
    "C0330", # Bad indentation
    "C0326", # Bad whitespace
    "W0511", # TODOs are acceptable
    "W0212", # Protected member access
    "W0703", # Broad except
    "W1201", # Logging format interpolation
    "W1202", # Logging format interpolation
    "C0412", # Ungrouped imports (handled by isort)
    "E1101", # No member (often false positives)
    "R0903", # Too few public methods
    "R0913", # Too many arguments
    "R0914", # Too many local variables
]

[tool.pylint.basic]
good-names = ["i", "j", "k", "ex", "Run", "_", "id", "db"]
class-naming-style = "PascalCase"
function-naming-style = "snake_case"
method-naming-style = "snake_case"
variable-naming-style = "snake_case"
const-naming-style = "UPPER_CASE"
argument-naming-style = "snake_case"
attr-naming-style = "snake_case"
module-naming-style = "snake_case"

[tool.pylint.format]
indent-string = "    "
indent-after-paren = 4
max-line-length = 100

[tool.pylint."messages control"]
confidence = ["HIGH", "CONTROL_FLOW", "INFERENCE", "INFERENCE_FAILURE", "UNDEFINED"]

[tool.pylint.design]
max-args = 8
max-attributes = 12
max-branches = 15
max-locals = 20
max-returns = 8
max-statements = 60
min-public-methods = 1
max-public-methods = 25

[tool.pylint.similarities]
min-similarity-lines = 6
ignore-comments = true
ignore-docstrings = true
ignore-imports = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []
extend-ignore = [
    "F403",  # 'from module import *' used; unable to detect undefined names
    "F405",  # 'name' may be undefined, or defined from star imports: module
    "F811",  # Redefinition of unused 'name' from line N
    "F841",  # Local variable 'name' is assigned to but never used
    # Disable some pydocstyle rules if needed
    "D203",  # 1 blank line required before class docstring
    "D212",  # Multi-line docstring summary should start at the first line
    "D213",  # Multi-line docstring summary should start at the second line
    "D404",  # First word of the docstring should not be 'This'
    "D415",  # First line should end with a period
    "D417",  # Missing argument description in the docstring
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in __init__
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"] # Allow asserts in tests
"*/__init__.py" = ["F401"] # Allow unused imports in __init__.py

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.isort]
known-first-party = ["app"]
